Adds memory check, fat and bmp headers
BMP (bitmap) and FAT(File Allocation Table) headers are added in this
commit. They'll form the base for out filesystem implementation. As
no real tests are added for these interfaces, no real code for them
as well.

This commit also includes memory checking for all tests, which will
be a condition (tests passing) for merging.